Parse class static blocks
Summary:
Class static blocks can't have `return` in them, so we have to check for
the Return param and error in the parser, instead of waiting for
SemanticValidator to error (where we'd have to track whether we are
currently in the top-level of a class static block).

We also have to forbid `await` expressions, which is easy to do by using
the `forbidAwaitExpression_` flag in `SemanticValidator`.

The ESTree spec adds a StaticBlock node which is a BlockStatement.

Reviewed By: fbmal7

Differential Revision: D65822124

fbshipit-source-id: 2c1f6989a42e07764fd1eaf639620131f068b71b