Add transform for match statements
Summary:
This builds on the previous diff implementing the transform for match expressions. See that for more details.

Output:
We replace the match statement with a labelled block containing a series of if statements. We use the label to exit this block when we have a successful match. Otherwise, we go to the next test.

Reviewed By: SamChou19815

Differential Revision: D66852555

fbshipit-source-id: 6c5459732b3d813afd94b0b39a7fee11dc017292