Add transform for match expressions
Summary:
Adds a transform for match expressions. Support for match statements, and "or" patterns with nested bindings will be added at a later point.

This is a complicated transform. I've added several builders and utilities to make it easier to implement. These can be used by other transforms in the future.

Steps:
- Analyze the case patterns and produce an abstract representation of them (arrays of `Condition`s and `Binding`s)
- If no bindings are introduced, produce a series of nested `ConditionalExpression`s (e.g. `c ? x : y`).
  - This way, we have a 0 overhead option for current patterns that don't introduce bindings
- If there are bindings, produce an immediately-invoked arrow function expression
  - Each case turns into an `if` statement
- If there is no wildcard case provided by the match, add a `throw` with an error as the final output

Reviewed By: SamChou19815

Differential Revision: D66835724

fbshipit-source-id: c42eba946afecd61f855495c7ff7ad1684aadd12