Pipe information through compiler for simplified resolver generation for property lookup resolvers
Summary:
@public
This diff creates the foundation of a new quality of life feature that allows for quickly defining resolvers that just do a property lookup on the base model. The changes in this diff are just piping the information of whether or not the resolver is a regular function or a new property lookup resolver through the compiler. It starts in the schema generation module and must carry the information all the way through to the printer. A stacked diff will actually implement the logic to find and add the resolver information.

Reviewed By: captbaritone

Differential Revision: D66714731

fbshipit-source-id: 4eba405cc590f461339f3f165d9832b52854c75e